requirements-dev.txt
pytest
mypy
isort
autoflake
flake8
black
pytest-coverage
tox
mock
mathy_pydoc>=0.7.22,<0.8.0
typer-cli

# test serializing spacy models using Pathy in place of Path
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.5/en_core_web_sm-2.2.5.tar.gz#egg=en_core_web_sm
spacy
